last update for 0615

+++ server/database/tables.py
@@ -379,7 +379,8 @@ class InTag(BASE):
 
     stockIn_alpha = Column(String(4))  # 入庫標籤上的批次文字, 2022/11/15討論增加
 
-    count_inv_modify = Column(Integer, default=0)  # 盤點數
+    #count_inv_modify = Column(Integer, default=0)  # 盤點數
+    count_inv_modify = Column(Float, default=0.0)  # 盤點數 2023-06-12 modify
 
     comment = Column(String(80), default='')  # 盤點說明  ,  , 11/26建議 default=''
